feat(falcon-script-watch): cleanup
